Refactored warnings into a hook used on <ErrorMessagesFragment>
