--- src/components/pages/deploy/atoms.ts
@@ -1,7 +1,0 @@
-import { atom } from "recoil";
-import { ValidationErrors } from "validators/types";
-
-export const warningsAtom = atom<ValidationErrors>({
-  key: "warnings",
-  default: {},
-});
