+++ src/components/pages/deploy/useWarnings.ts
@@ -1,27 +1,18 @@
-import React, { memo, useEffect } from "react";
-import { useSetRecoilState } from "recoil";
-import { useForm, useFormState } from "react-final-form";
-import Decimal from "decimal.js";
+import { useEffect, useMemo } from "react";
+import { useForm } from "react-final-form";
 import { capitalize } from "lodash";
 
-import { setFieldData } from "utils/finalForm";
 import {
-  Network,
   FUNDING_PER_BRACKET_WARNING_THRESHOLD,
-  ONE_HUNDRED_DECIMAL,
   START_PRICE_WARNING_THRESHOLD_PERCENTAGE,
 } from "utils/constants";
 
-import { useGetPrice } from "hooks/useGetPrice";
-import { useTokenDetails } from "hooks/useTokenDetails";
-import { useAmountInUsd } from "hooks/useAmountInUsd";
-import { useSafeInfo } from "hooks/useSafeInfo";
+import { ValidationError, ValidationErrors } from "validators/types";
 
-import { ValidationError } from "validators/types";
-
-import { warningsAtom } from "./atoms";
 import { DeployFormValues } from "./types";
 import { getBracketValue } from "./DeployForm";
+import { useFundingOutOfThreshold } from "./useFundingOutOfThreshold";
+import { useIsStartPriceOutOfThreshold } from "./useStartPriceOutOfThreshold";
 
 const TOTAL_BRACKETS_WARNING = {
   label: "Strategy with only one bracket",
@@ -40,6 +31,7 @@ function buildStartPriceWarning(threshold?: number): ValidationError {
   const direction = threshold > 0 ? "higher" : "lower";
   const label = `Detected: Start price >2% ${direction} than market price`;
   const children = `The specified Start Price is at least 2% ${direction} than the current indicated market price. If intentional continue with your strategy.`;
+
   return { label, children };
 }
 
@@ -56,98 +48,15 @@ function buildPriceWarning(type: "lowest" | "highest"): ValidationError {
   };
 }
 
-type UseIsStartPriceOutOfThresholdParams = {
-  startPrice?: string;
-  threshold: number;
-  baseTokenAddress?: string;
-  quoteTokenAddress?: string;
-};
-
 /**
- * Checks whether given `startPrice` is out of given `threshold` by comparing with market price.
- * Returns undefined on invalid input.
- * Returns 0 when `startPrice` is within threshold.
- * Returns +1 when `startPrice` is above threshold.
- * Returns -1 when `startPrice` is below threshold.
+ * Hook that goes through form values and returns warnings, if any.
+ * It also sets fields `warn` flag.
+ *
+ * Must be used within the Form context
+ *
+ * @param values Form values
  */
-function useIsStartPriceOutOfThreshold(
-  params: UseIsStartPriceOutOfThresholdParams
-): number | undefined {
-  const { baseTokenAddress, quoteTokenAddress, startPrice, threshold } = params;
-
-  // Avoid additional queries if `startPrice` is not set
-  const baseToken = useTokenDetails(startPrice ? baseTokenAddress : undefined);
-  const quoteToken = useTokenDetails(
-    startPrice ? quoteTokenAddress : undefined
-  );
-
-  const { price } = useGetPrice({ baseToken, quoteToken });
-
-  if (!startPrice || isNaN(+startPrice) || !price) {
-    return undefined;
-  }
-
-  const startPriceDecimal = new Decimal(startPrice);
-
-  const priceDifferencePercentage = startPriceDecimal
-    .minus(price)
-    .div(price)
-    .mul(ONE_HUNDRED_DECIMAL);
-
-  const thresholdDecimal = new Decimal(threshold);
-
-  // negative threshold --- zero ----- positive threshold
-  // ########|----------------|----------------|#########
-  //    -1   |                0                |   +1
-  return priceDifferencePercentage.gt(thresholdDecimal)
-    ? 1
-    : !priceDifferencePercentage.gte(thresholdDecimal.negated())
-    ? -1
-    : 0;
-}
-
-type UseFundingOutOfThresholdParams = {
-  fundingAmount?: string;
-  brackets?: number;
-  tokenAddress?: string;
-  threshold: number;
-};
-
-function useFundingOutOfThreshold(
-  params: UseFundingOutOfThresholdParams
-): boolean {
-  const { fundingAmount, brackets, tokenAddress, threshold } = params;
-
-  const { amountInUsd } = useAmountInUsd({
-    tokenAddress,
-    amount: fundingAmount,
-  });
-
-  const { network } = useSafeInfo();
-
-  if (
-    !fundingAmount ||
-    isNaN(+fundingAmount) ||
-    !brackets ||
-    !amountInUsd ||
-    Network[network] === Network.xdai
-  ) {
-    return false;
-  }
-
-  return !amountInUsd.div(brackets).gte(threshold);
-}
-
-export type Props = {
-  mutators: { setFieldData: typeof setFieldData };
-  values: DeployFormValues;
-};
-
-export const Warnings = memo(function Warnings(): JSX.Element {
-  const setWarnings = useSetRecoilState(warningsAtom);
-  const { values } = useFormState<DeployFormValues>({
-    subscription: { values: true },
-  });
+export function useWarnings(values: DeployFormValues): ValidationErrors {
   const {
     mutators: { setFieldData },
   } = useForm();
@@ -201,11 +110,19 @@ export const Warnings = memo(function Warnings(): JSX.Element {
     setFieldData("startPrice", { warn: startPriceHasWarning });
     setFieldData("lowestPrice", { warn: lowestPriceHasWarning });
     setFieldData("highestPrice", { warn: highestPriceHasWarning });
+  }, [
+    baseTokenAmountHasWarning,
+    highestPriceHasWarning,
+    lowestPriceHasWarning,
+    quoteTokenAmountHasWarning,
+    setFieldData,
+    startPriceHasWarning,
+    totalBracketsHasWarning,
+  ]);
 
-    // 2. This part aggregates all the warning messages to display them at the end of the form.
-    // Since we can't easily access the field `data` state, we are using an external source
-    setWarnings((warnings) => ({
-      ...warnings,
+  // 2. This part aggregates all the warning messages to display them at the end of the form.
+  const warnings = useMemo(
+    () => ({
       totalBrackets: totalBracketsHasWarning && TOTAL_BRACKETS_WARNING,
       tokenAmount:
         (baseTokenAmountHasWarning || quoteTokenAmountHasWarning) &&
@@ -215,10 +132,17 @@ export const Warnings = memo(function Warnings(): JSX.Element {
         buildStartPriceWarning(startPriceOutOfThreshold),
       lowestPrice: lowestPriceHasWarning && buildPriceWarning("lowest"),
       highestPrice: highestPriceHasWarning && buildPriceWarning("highest"),
-    }));
-  }, [baseTokenAmountHasWarning, highestPriceHasWarning, lowestPriceHasWarning, quoteTokenAmountHasWarning, setFieldData, setWarnings, startPriceHasWarning, startPriceOutOfThreshold, totalBracketsHasWarning]);
+    }),
+    [
+      baseTokenAmountHasWarning,
+      highestPriceHasWarning,
+      lowestPriceHasWarning,
+      quoteTokenAmountHasWarning,
+      startPriceHasWarning,
+      startPriceOutOfThreshold,
+      totalBracketsHasWarning,
+    ]
+  );
 
-  // Nothing to render here, we just wanna update the form state.
-  // TODO: maybe this should be moved to <ErrorMessagesFragment/>?
-  return <></>;
-});
+  return warnings;
+}
